move simple token parser tests
